Prevent vtkDebugLeaks from polluting process output if no leaks are found
When VTK was built with VTK_DEBUG_LEAKS enabled then always printed
"vtkDebugLeaks has found no leaks." message on the process error output
when everything was fine, which broke features that relied on passing information
via process outputs.

This informational message is now logged at trace level, so that the application
can decide if it gets printed to the process output or logged to file.