Fix crash in vtkOpenGLFramebufferObject::GetMultiSamples()
this->ColorBuffers[abuff] created a new map entry when the active buffer was not found.
Since this entry contained nullptr for vtkFOInfo*, it caused a crash anytime when vtkFOInfo* pointer was attempted to be used.

Fixed by finding the map element instead of looking up using operator[] (the same way as it is done everywhere else in the file).