[Backport MR-10783] VR: Move the TrackHMD flag to the camera
Track HMD affects whether the camera is tracking the head mounted
display or not. Moving this flag to the vtkVRCamera is fitting for this
purpose to allow multiple cameras in the VR world.  When performing
hardware picking, the camera is moved to the wand and TrackHMD is
disabled to ensure that the scene doesn't change.

Merge-request: vtk/vtk!10783