[Backport] Fix bounds computation in vtkMultiVolume
This line set the homogeneous IJK coordinates incorrectly:

    pointsDataCoords.emplace_back(bounds[1], bounds[3], bounds[5], 0.);

The problem is that the last component is set to 0 instead of 1
(the coordinates specify a direction vector and not a position).
When this vector is multiplied by the homogeneous transformation matrix,
the position is incorrect, because the translation component of the matrix
is not taken into account.

fixes https://github.com/Slicer/Slicer/issues/8301
fixes https://gitlab.kitware.com/vtk/vtk/-/issues/19447

(cherry picked from commit 5e878068c23f6b87a398dbe8000424971dcb7afa)